split adminprivate into bukkit and shared
