+++ instances/bukkit/build.gradle.kts
@@ -38,7 +38,8 @@ dependencies {
     implementation(projects.modules.coreBukkit)
     implementation(projects.modules.menu)
     implementation(projects.modules.broadcast)
-    implementation(projects.modules.adminprivate)
+    implementation(projects.modules.adminprivate.api)
+    implementation(projects.modules.adminprivate.bukkit)
     implementation(projects.modules.discordlink)
     implementation(projects.modules.townyDiscord)
     implementation(projects.modules.moneydrop)
