+++ modules/adminprivate/api/build.gradle
@@ -1,0 +1,21 @@
+plugins {
+    id("org.jetbrains.kotlin.jvm")
+    id("org.jetbrains.kotlin.plugin.serialization")
+}
+
+dependencies {
+    // Kotlin
+    implementation(libs.bundles.kotlin)
+    // AstraLibs
+    implementation(libs.minecraft.astralibs.core)
+    implementation(libs.minecraft.astralibs.command)
+    implementation(libs.klibs.kstorage)
+    implementation(libs.klibs.mikro.core)
+    // Test
+    testImplementation(libs.bundles.testing.kotlin)
+    testImplementation(libs.tests.kotlin.test)
+    // Local
+    implementation(projects.modules.core)
+
+    compileOnly("com.google.guava:guava:31.1-jre")
+}
