+++ modules/adminprivate/api/src/main/kotlin/ru/astrainteractive/aspekt/module/adminprivate/data/krate/AdminPrivateKrate.kt
