+++ modules/adminprivate/api/src/main/kotlin/ru/astrainteractive/aspekt/module/adminprivate/debounce/RetractKey.kt
@@ -1,0 +1,5 @@
+package ru.astrainteractive.aspekt.module.adminprivate.debounce
+
+interface RetractKey {
+    class Vararg(vararg val value: Any) : RetractKey
+}
