+++ modules/adminprivate/api/src/main/kotlin/ru/astrainteractive/aspekt/module/adminprivate/event/SharedCancellableEvent.kt
@@ -1,0 +1,5 @@
+package ru.astrainteractive.aspekt.module.adminprivate.event
+
+interface SharedCancellableEvent {
+    var isCancelled: Boolean
+}
