+++ modules/adminprivate/api/src/main/kotlin/ru/astrainteractive/aspekt/module/adminprivate/model/AdminChunk.kt
@@ -1,0 +1,12 @@
+package ru.astrainteractive.aspekt.module.adminprivate.model
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class AdminChunk(
+    val x: Int,
+    val z: Int,
+    val worldName: String,
+    val flags: Map<ChunkFlag, Boolean>,
+    val chunkKey: Long,
+)
