+++ modules/adminprivate/api/src/main/kotlin/ru/astrainteractive/aspekt/module/adminprivate/model/AdminPrivateConfig.kt
@@ -1,0 +1,9 @@
+package ru.astrainteractive.aspekt.module.adminprivate.model
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class AdminPrivateConfig(
+    val chunks: Map<String, AdminChunk> = emptyMap(),
+    val isEnabled: Boolean = true
+)
