+++ modules/adminprivate/api/src/main/kotlin/ru/astrainteractive/aspekt/module/adminprivate/model/ChunkFlag.kt
@@ -1,0 +1,16 @@
+package ru.astrainteractive.aspekt.module.adminprivate.model
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+enum class ChunkFlag {
+    BREAK,
+    PLACE,
+    INTERACT,
+    EXPLODE,
+    EMPTY_BUCKET,
+    SPREAD,
+    RECEIVE_DAMAGE,
+    HOSTILE_MOB_SPAWN,
+    ICE_MELT
+}
