--- modules/adminprivate/build.gradle
@@ -1,25 +1,0 @@
-plugins {
-    id("org.jetbrains.kotlin.jvm")
-    id("org.jetbrains.kotlin.plugin.serialization")
-}
-
-dependencies {
-    // Kotlin
-    implementation(libs.bundles.kotlin)
-    // Spigot dependencies
-    compileOnly(libs.minecraft.paper.api)
-    // AstraLibs
-    implementation(libs.minecraft.astralibs.core)
-    implementation(libs.minecraft.astralibs.command)
-    implementation(libs.minecraft.astralibs.command.bukkit)
-    implementation(libs.klibs.kstorage)
-    implementation(libs.klibs.mikro.core)
-    implementation(libs.minecraft.astralibs.menu.bukkit)
-    implementation(libs.minecraft.astralibs.core.bukkit)
-    // Test
-    testImplementation(libs.bundles.testing.kotlin)
-    testImplementation(libs.tests.kotlin.test)
-    // Local
-    implementation(projects.modules.core)
-    implementation(projects.modules.coreBukkit)
-}
