--- modules/adminprivate/src/main/kotlin/ru/astrainteractive/aspekt/module/adminprivate/controller/di/AdminPrivateControllerDependencies.kt
@@ -1,22 +1,0 @@
-package ru.astrainteractive.aspekt.module.adminprivate.controller.di
-
-import ru.astrainteractive.aspekt.di.CoreModule
-import ru.astrainteractive.aspekt.module.adminprivate.data.AdminPrivateRepository
-import ru.astrainteractive.aspekt.module.adminprivate.data.AdminPrivateRepositoryImpl
-import ru.astrainteractive.aspekt.module.adminprivate.di.AdminPrivateModule
-
-internal interface AdminPrivateControllerDependencies {
-    val repository: AdminPrivateRepository
-
-    class Default(
-        coreModule: CoreModule,
-        adminPrivateModule: AdminPrivateModule
-    ) : AdminPrivateControllerDependencies {
-        override val repository: AdminPrivateRepository by lazy {
-            AdminPrivateRepositoryImpl(
-                file = adminPrivateModule.adminChunksFile,
-                stringFormat = coreModule.yamlFormat
-            )
-        }
-    }
-}
