--- modules/adminprivate/src/main/kotlin/ru/astrainteractive/aspekt/module/adminprivate/debounce/RetractKey.kt
@@ -1,5 +1,0 @@
-package ru.astrainteractive.aspekt.module.adminprivate.debounce
-
-internal interface RetractKey {
-    class Vararg(vararg val value: Any) : RetractKey
-}
