--- modules/adminprivate/src/main/kotlin/ru/astrainteractive/aspekt/module/adminprivate/model/AdminChunk.kt
@@ -1,12 +1,0 @@
-package ru.astrainteractive.aspekt.module.adminprivate.model
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-internal data class AdminChunk(
-    val x: Int,
-    val z: Int,
-    val worldName: String,
-    val flags: Map<ChunkFlag, Boolean>,
-    val chunkKey: Long,
-)
