--- modules/adminprivate/src/main/kotlin/ru/astrainteractive/aspekt/module/adminprivate/model/AdminPrivateConfig.kt
@@ -1,9 +1,0 @@
-package ru.astrainteractive.aspekt.module.adminprivate.model
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-internal data class AdminPrivateConfig(
-    val chunks: Map<String, AdminChunk> = emptyMap(),
-    val isEnabled: Boolean = true
-)
